Updating creation flow unit test location
