+++ Composer/packages/client/__tests__/components/CreationFlow/DefineConversation/index.test.tsx
@@ -9,7 +9,6 @@ import { StorageFolder } from '../../../../src/recoilModel/types';
 import { focusedStorageFolderState, storagesState } from '../../../../src/recoilModel';
 import DefineConversation from '../../../../src/components/CreationFlow/DefineConversation';
 
-// TODO change to new creation component
 describe('<DefineConversation/>', () => {
   const onCurrentPathUpdateMock = jest.fn();
   const onSubmitMock = jest.fn();
@@ -55,34 +54,6 @@ describe('<DefineConversation/>', () => {
   it('should render the component', () => {
     const component = renderComponent();
     expect(component.container).toBeDefined();
-  });
-
-  it('should update formdata with data passed through location props', async () => {
-    locationMock = {
-      search:
-        'schemaUrl%3Dhttps%3A%2F%2Fraw.githubusercontent.com%2Fmicrosoft%2Fbotframework-sdk%2Fmaster%2Fschemas%2Fcomponent%2Fcomponent.schema%26name%3DEchoBot-11299%26description%3DTest%20Echo',
-    };
-    const component = renderComponent();
-    const node = await component.findByText('OK');
-    fireEvent.click(node);
-    expect(
-      onSubmitMock.mock.calls[0][0] ===
-        {
-          description: 'Test Echo',
-          name: 'EchoBot-11299',
-          location: '\\test-folder\\Desktop',
-          schemaUrl:
-            'https://raw.githubusercontent.com/microsoft/botframework-sdk/master/schemas/component/component.schema',
-        } ||
-        onSubmitMock.mock.calls[0][0] ===
-          {
-            description: 'Test Echo',
-            name: 'EchoBot-11299',
-            location: '/test-folder/Desktop',
-            schemaUrl:
-              'https://raw.githubusercontent.com/microsoft/botframework-sdk/master/schemas/component/component.schema',
-          }
-    ).toBeTruthy;
   });
 
   it('does not allow submission when the name is invalid', async () => {
